Refactor domain to core and inline i18n types
Moved files from 'domain' to 'core' for better organization and updated imports accordingly. Inlined Language and Translations types into useI18n.ts, removing the separate i18n.ts type file.
