+++ src/app/components/TomieTerminal/utils/moodStyleGenerator.ts
@@ -1,5 +1,5 @@
 import { Mood } from '../../../types/mood';
-import { moodColors } from '../../../domain/moodConfig';
+import { moodColors } from '../../../core/moodConfig';
 
 export const generateMoodStyles = (currentMood: Mood): string => {
     const currentColors = moodColors[currentMood];
