+++ src/app/core/moodConfig.ts
