+++ src/app/core/stateManager.ts
